+++ mycodo/config.py
@@ -14,7 +14,7 @@ sys.path.append(os.path.abspath(os.path.dirname(__file__)))
 from config_translations import TRANSLATIONS
 
 MYCODO_VERSION = '8.6.4'
-ALEMBIC_VERSION = '4d3258ef5864'
+ALEMBIC_VERSION = 'd66e33093e8e'
 
 #  FORCE_UPGRADE_MASTER
 #  Set True to enable upgrading to the master branch of the Mycodo repository.
@@ -57,6 +57,7 @@ DASHBOARD_WIDGETS = [
     ('output_pwm_slider', '{}: {}'.format(
         TRANSLATIONS['output']['title'], lazy_gettext('PWM Slider'))),
     ('pid_control', lazy_gettext('PID Control')),
+    ('python_code', lazy_gettext('Python Code')),
     ('camera', TRANSLATIONS['camera']['title'])
 ]